+++ backend/dotenv
@@ -2,3 +2,5 @@ DB_HOST = localhost
 DB_USER = root
 DB_PASSWORD = **YOUR PASSWORD HERE!!!**
 DB_NAME = dio_membership_test
+
+PORT = 3001
