+++ backend/package.json
@@ -8,6 +8,6 @@
     "mysql2": "^3.12.0"
   },
   "scripts": {
-    "start": "node index.js"
+    "start": "node server.js"
   }
 }
