+++ backend/server.js
@@ -1,0 +1,8 @@
+require('dotenv').config({path: '.env'});
+const app = require("./src/app");
+
+const PORT = process.env.PORT || 3001;
+
+app.listen(PORT, () => {
+    console.log(`Server listening on ${PORT}`);
+});
