+++ backend/src/app.js
@@ -2,8 +2,6 @@ const express = require("express");
 const cors = require('cors');
 const { upload, handleFileUpload } = require('./handleFileUpload');
 
-const PORT = process.env.PORT || 3001;
-
 const app = express();
 app.use(cors());
 
@@ -14,6 +12,6 @@ app.get('/health', (req, res) => {
 
 app.post('/upload', upload.single('csv_file'), handleFileUpload);
 
-app.listen(PORT, () => {
-    console.log(`Server listening on ${PORT}`);
-});
+// TODO: Add more REST endpoints here
+
+module.exports = app;
