+++ backend/src/handleFileUpload.js
@@ -1,7 +1,7 @@
 const multer = require('multer');
 const path = require('path');
 const fs = require('fs');
-const csvProcessor = require('./load_csv');
+const csvProcessor = require('./loadCsv');
 
 const storage = multer.diskStorage({
     destination: (req, file, cb) => {
